+++ src/main/java/com/c9Pay/creditservice/web/exception/handler/ExceptionResponse.java
@@ -1,0 +1,16 @@
+package com.c9Pay.creditservice.web.exception.handler;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+import java.util.Date;
+
+@Data
+@AllArgsConstructor
+public class ExceptionResponse {
+
+    private Date timeStamp;
+    private String message;
+    private String details;
+
+}
