+++ src/main/java/com/c9Pay/creditservice/web/exception/handler/GlobalExceptionHandler.java
@@ -1,0 +1,23 @@
+package com.c9Pay.creditservice.web.exception.handler;
+
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpStatusCode;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.context.request.WebRequest;
+import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
+
+import java.util.Date;
+
+@ControllerAdvice
+public class GlobalExceptionHandler extends ResponseEntityExceptionHandler {
+    @ExceptionHandler
+    public ResponseEntity<?> handlerAllException(Exception e, WebRequest request){
+        ExceptionResponse exceptionResponse =
+                new ExceptionResponse(new Date(),e.getMessage(), request.getDescription(false));
+
+        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(exceptionResponse);
+    }
+}
