+++ src/main/java/com/c9Pay/creditservice/web/repository/AccountRepository.java
@@ -1,4 +1,4 @@
-package com.c9Pay.creditservice.repository;
+package com.c9Pay.creditservice.web.repository;
 import java.util.List;
 import java.util.Optional;
